pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "divoom-client"
version = "0.1.0"
description = "A display manager for Divoom Pixoo 64 LED displays"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Jason Person", email = "jsperson@gmail.com" }
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "typer>=0.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "apscheduler>=3.10.0",
    "yfinance>=0.2.0",
    "requests>=2.31.0",
    "Pillow>=10.0.0",
    "jsonpath-ng>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
divoom = "divoom_client.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/divoom_client"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
